// src/models/flatexRecord.ts
export class FlatexRecord {
    bookingDate: string;
    valueDate: string;
    description: string;
    isin: string;
    quantity: number;
    unit: string;
    amount: number;
    currency: string;
    price: number;
    priceCurrency: string;
    exchangeRate: number;
    transactionNumber: string;
    transactionInformation: string;
}